fixed database types

+++ app/models.py
@@ -3,5 +3,5 @@ from .extensions import db
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     phone = db.Column(db.String(50), nullable=False)
-    longitude = db.Column(db.Integer)
-    latitude = db.Column(db.Integer)
+    longitude = db.Column(db.Float)
+    latitude = db.Column(db.Float)
